つくチャレのみの不要ファイルを整理、最新化
