--- map/map_white.yaml
@@ -1,7 +1,0 @@
-image: map_white.pgm
-mode: trinary
-resolution: 0.1
-origin: [-250.0, -250.0, 0]
-negate: 0
-occupied_thresh: 0.65
-free_thresh: 0.25
