+++ map/put_map_here
@@ -1,0 +1,1 @@
+This directory is where you will save your maps
